Update Envoy to f84440dc4f95890f14e2e0686b07258f030b54b3 (#321)
- Trace id generation broke. This was amended by directly
using the new request id uid implementation. It would be
nice to leverage the new request id extension capability
later for this.
- Amended calls to get counters by name
- Synced .bazelrc
- Histogram value formatting subtly changed in some cases, probably caused by an update
  to Envoy's dependency on `libfmt`. 
  Introduce an explicit formatting specifier in the affected output formatting  code to address that.
- Tweaks for resolving new OOM issues with ASAN/TSAN in CI that started with the updated Envoy 
   revision. This affected both the build process as well as the testing process itself. Notably, the
   python-based integration tests would get OOM-killed in asan runs.

+++ source/client/output_formatter_impl.cc
@@ -91,8 +91,9 @@ std::string ConsoleOutputFormatterImpl::formatProto(const nighthawk::client::Out
               header_written = true;
             }
             ss << fmt::format("  {:<{}}{:<{}}{:<{}}", p, 12, percentile.count(), 12,
-                              percentile.has_duration() ? formatProtoDuration(percentile.duration())
-                                                        : fmt::format("{}", percentile.raw_value()),
+                              percentile.has_duration()
+                                  ? formatProtoDuration(percentile.duration())
+                                  : fmt::format("{}", static_cast<int64_t>(percentile.raw_value())),
                               15)
                << std::endl;
           }
@@ -189,7 +190,8 @@ DottedStringOutputFormatterImpl::formatProto(const nighthawk::client::Output& ou
               "{}.microseconds: {}", percentile_prefix,
               Envoy::Protobuf::util::TimeUtil::DurationToMicroseconds(percentile.duration()));
         } else {
-          ss << fmt::format("{}.value: {}", percentile_prefix, percentile.raw_value());
+          ss << fmt::format("{}.value: {}", percentile_prefix,
+                            static_cast<int64_t>(percentile.raw_value()));
         }
         ss << std::endl;
       });
